Interior_AI/backend/requirements.txt
Flask
Flask-Cors
Pillow
requests
python-dotenv
